backend/requirements.txt
# ══════════════════════════════════════════════════════════════
# Research Paper Assistant — Backend Requirements
# ══════════════════════════════════════════════════════════════

# ── Framework & Server ────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ── Database & ORM ────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# ── Redis ─────────────────────────────────────────────────────
redis[hiredis]==5.0.1

# ── Vector Database ───────────────────────────────────────────
qdrant-client==1.7.0

# ── LLM & AI ─────────────────────────────────────────────────
litellm==1.30.0
langchain==0.1.6
langchain-community==0.0.16

# ── AI Agents ─────────────────────────────────────────────────
crewai==0.114.0

# ── ML Models ─────────────────────────────────────────────────
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
sentencepiece==0.1.99

# ── NLP & Text Processing ────────────────────────────────────
nltk==3.8.1
tiktoken==0.5.2

# ── OCR ───────────────────────────────────────────────────────
paddleocr==2.7.3
paddlepaddle==2.6.0

# ── Voice — STT ───────────────────────────────────────────────
faster-whisper==1.0.1

# ── Voice — TTS ───────────────────────────────────────────────
edge-tts==6.1.9

# ── PDF Processing ────────────────────────────────────────────
PyPDF2==3.0.1
pdfplumber==0.10.3
fpdf2==2.7.8

# ── DOCX Processing ──────────────────────────────────────────
python-docx==1.1.0

# ── Image Processing ─────────────────────────────────────────
Pillow==10.2.0

# ── File Security ────────────────────────────────────────────
python-magic-bin==0.4.14

# ── HTTP Client ───────────────────────────────────────────────
httpx==0.26.0

# ── Rate Limiting ─────────────────────────────────────────────
slowapi==0.1.9

# ── Logging ───────────────────────────────────────────────────
structlog==24.1.0
loguru==0.7.2

# ── Google Gemini (for STT/TTS) ───────────────────────────────
google-generativeai==0.3.2

# ── ElevenLabs ────────────────────────────────────────────────
elevenlabs==1.0.0

# ── WebSocket ─────────────────────────────────────────────────
websockets==12.0

# ── Utilities ─────────────────────────────────────────────────
python-multipart==0.0.6
aiofiles==23.2.1
